Fixes to work with chrome storage

+++ assets/js/preview.js
@@ -1,29 +1,34 @@
-if (!chrome || !chrome.storage) {
-    exit;
+function try_chrome_storage() {
+    if (!chrome || !chrome.storage || !chrome.storage.local) {
+        return;
+    }
+
+    chrome.storage.local.clear();
+
+    chrome.storage.onChanged.addListener(function (changes) {
+
+        if (changes.css) {
+            if (changes.css.newValue) {
+                document.querySelector("#nexus_prototype_styles").innerHTML = changes.css.newValue;
+            }
+        }
+
+        if (changes.html) {
+            if (changes.html.newValue) {
+                document.querySelector("body").innerHTML = changes.html.newValue;
+            }
+        }
+
+        if (changes.js) {
+            if (changes.js.newValue) {
+                document.querySelector("#nexus_prototype_scripts").innerHTML = changes.js.newValue;
+                console.debug(changes.js.newValue);
+                eval(changes.js.newValue);
+            }
+        }
+
+    });
+
 }
 
-chrome.storage.local.clear();
-
-chrome.storage.onChanged.addListener(function (changes) {
-
-    if (changes.css) {
-        if (changes.css.newValue) {
-            document.querySelector("#nexus_prototype_styles").innerHTML = changes.css.newValue;
-        }
-    }
-
-    if (changes.html) {
-        if (changes.html.newValue) {
-            document.querySelector("body").innerHTML = changes.html.newValue;
-        }
-    }
-
-    if (changes.js) {
-        if (changes.js.newValue) {
-            document.querySelector("#nexus_prototype_scripts").innerHTML = changes.js.newValue;
-            console.debug(changes.js.newValue);
-            eval(changes.js.newValue);
-        }
-    }
-
-});
+try_chrome_storage();
